requirements.txt
fastapi==0.118.0
fastmcp==2.12.4
chromadb==1.1.0
httpx==0.28.1
pydantic==2.11.10
uvicorn==0.37.0
python-dotenv==1.1.1
sse-starlette==3.0.2
sentence-transformers==3.3.1
